Makefile and bugfix
A Makefile has been created that will run the python script
every time and the rest if new data was retrieved.

Also, it seems possible that new data from coinmetrics can be
incomplete which yiels an error. If either the USD price or
number of coins is null, the script exits.